refactor(rolldown_plugin_reporter): move utility logic to utils.rs and align plugin name
