+++ Cargo.toml
@@ -107,7 +107,7 @@ rolldown_plugin_manifest = { version = "0.1.0", path = "./crates/rolldown_plugin
 rolldown_plugin_module_federation = { version = "0.1.0", path = "./crates/rolldown_plugin_module_federation" }
 rolldown_plugin_module_preload_polyfill = { version = "0.1.0", path = "./crates/rolldown_plugin_module_preload_polyfill" }
 rolldown_plugin_replace = { version = "0.1.0", path = "./crates/rolldown_plugin_replace" }
-rolldown_plugin_report = { version = "0.1.0", path = "./crates/rolldown_plugin_report" }
+rolldown_plugin_reporter = { version = "0.1.0", path = "./crates/rolldown_plugin_reporter" }
 rolldown_plugin_transform = { version = "0.1.0", path = "./crates/rolldown_plugin_transform" }
 rolldown_plugin_vite_resolve = { version = "0.1.0", path = "./crates/rolldown_plugin_vite_resolve" }
 rolldown_plugin_wasm_fallback = { version = "0.1.0", path = "./crates/rolldown_plugin_wasm_fallback" }
