+++ crates/rolldown_plugin_reporter/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rolldown_plugin_report"
+name = "rolldown_plugin_reporter"
 version = "0.1.0"
 edition.workspace = true
 homepage.workspace = true
