+++ crates/rolldown_plugin_reporter/src/lib.rs
@@ -1,6 +1,8 @@
+mod utils;
+
 use std::{
   borrow::Cow,
-  io::{StdoutLock, Write},
+  io::Write,
   path::Path,
   sync::{
     Arc, RwLock,
@@ -13,14 +15,14 @@ use rolldown_plugin::{HookUsage, Plugin, PluginContext};
 use sugar_path::SugarPath;
 
 #[derive(Debug)]
-pub struct ReportPlugin {
+pub struct ReporterPlugin {
   latest_checkpoint: Arc<RwLock<Instant>>,
   count: AtomicU32,
   has_transformed: AtomicBool,
   is_tty: bool,
 }
 
-impl ReportPlugin {
+impl ReporterPlugin {
   pub fn new(is_tty: bool) -> Self {
     Self {
       latest_checkpoint: Arc::new(RwLock::new(Instant::now())),
@@ -31,26 +33,9 @@ impl ReportPlugin {
   }
 }
 
-#[inline]
-#[allow(clippy::print_stdout)]
-pub fn clear_current_line() -> StdoutLock<'static> {
-  let mut lock = std::io::stdout().lock();
-  write!(&mut lock, "\x1B[2K\r").unwrap(); // clear current line and move cursor to the beginning
-  lock.flush().unwrap();
-  lock
-}
-
-#[inline]
-#[allow(clippy::print_stdout)]
-fn write_line(line: &str) {
-  let mut lock = clear_current_line();
-  write!(&mut lock, "{line}",).unwrap();
-  lock.flush().unwrap();
-}
-
-impl Plugin for ReportPlugin {
+impl Plugin for ReporterPlugin {
   fn name(&self) -> Cow<'static, str> {
-    Cow::Borrowed("builtin:report")
+    Cow::Borrowed("builtin:reporter")
   }
 
   async fn transform(
@@ -72,9 +57,13 @@ impl Plugin for ReportPlugin {
         }
         let relative_path = Path::new(args.id).relative(ctx.inner.cwd());
         // fetch_add return previous value
-        write_line(&format!("transforming ({}) {}", count + 1, relative_path.to_string_lossy()));
+        utils::write_line(&format!(
+          "transforming ({}) {}",
+          count + 1,
+          relative_path.to_string_lossy()
+        ));
       } else if !self.has_transformed.load(Ordering::Relaxed) {
-        write_line("transforming...");
+        utils::write_line("transforming...");
       }
       *self.latest_checkpoint.write().unwrap() = now;
     }
@@ -98,11 +87,11 @@ impl Plugin for ReportPlugin {
     _args: Option<&rolldown_plugin::HookBuildEndArgs<'_>>,
   ) -> rolldown_plugin::HookNoopReturn {
     if self.is_tty {
-      write_line("");
+      utils::write_line("");
     }
     let count = self.count.load(Ordering::SeqCst);
 
-    let mut lock = clear_current_line();
+    let mut lock = utils::clear_current_line();
     // print text with green fg color
     write!(&mut lock, "\x1b[32m{}\x1b[0m", "✓").unwrap();
     writeln!(&mut lock, " {count} modules transformed.",).unwrap();
